Refactor : DTO -> Entities
