+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/Auth/AuthService.java
@@ -5,8 +5,8 @@ import com.taskmanager.taskmanagerapi.Auth.authResponse.AuthResponse;
 import com.taskmanager.taskmanagerapi.Auth.authResponse.AuthResponseErr;
 import com.taskmanager.taskmanagerapi.Auth.authResponse.AuthResponseOk;
 import com.taskmanager.taskmanagerapi.Auth.authResponse.ErrorData;
-import com.taskmanager.taskmanagerapi.dto.User;
-import com.taskmanager.taskmanagerapi.dto.UserResponse;
+import com.taskmanager.taskmanagerapi.entities.User;
+import com.taskmanager.taskmanagerapi.entities.UserResponse;
 import com.taskmanager.taskmanagerapi.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
