+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/Auth/Jwt/JwtService.java
@@ -1,6 +1,5 @@
 package com.taskmanager.taskmanagerapi.Auth.Jwt;
 
-import com.taskmanager.taskmanagerapi.dto.User;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
