+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/Auth/authResponse/AuthResponse.java
@@ -1,13 +1,5 @@
 package com.taskmanager.taskmanagerapi.Auth.authResponse;
 
-import com.taskmanager.taskmanagerapi.dto.User;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Date;
-
 
 public abstract class AuthResponse{
