+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/Auth/authResponse/AuthResponseOk.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.Auth.authResponse;
 
-import com.taskmanager.taskmanagerapi.dto.User;
+import com.taskmanager.taskmanagerapi.entities.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
