+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/entities/Project.java
@@ -1,5 +1,4 @@
-package com.taskmanager.taskmanagerapi.dto;
-
+package com.taskmanager.taskmanagerapi.entities;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -8,49 +7,31 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.util.Date;
-
+import java.util.Set;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 @Entity
-@Table(name = "Task")
-public class Task {
+@Table(name="Projects")
+public class Project {
 
     @Id
     @GeneratedValue
     private int id;
 
-    private String title;
+    private String name;
 
     private String description;
 
-    private Date due_date;
 
-    private String priority;
-
-    private String category;
-
-    private String status;
-
-    // Todo : Verify if is required to use @PK notation
-
-    @OneToOne
-    private User created_by_user_id;
-
-    @OneToOne
-    private User assigned_to_user_id;
+    @OneToMany(fetch = FetchType.LAZY)
+    private Set<User> created_by_user_id;
 
     private Date created_at;
 
     private Date updated_at;
 
     private Date deleted_at;
-
-
-
-
-
-
 }
