+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/entities/Role.java
@@ -1,8 +1,8 @@
-package com.taskmanager.taskmanagerapi.dto;
+package com.taskmanager.taskmanagerapi.entities;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,31 +10,27 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.util.Date;
-import java.util.Set;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 @Entity
-@Table(name="Reports")
-public class Report {
+@Table(name = "Roles")
+public class Role {
 
     @Id
+    @GeneratedValue
     private int id;
 
-    @OneToMany
-    private Set<User> generated_by_user_id;
-
-    private Date generated_date;
-
-    // todo: Talk with the group of what datatype is spected, cause the "json" is not Sql standard
-    private String report_data;
+    private String role;
 
     private Date created_at;
 
     private Date updated_at;
 
     private Date deleted_at;
+
+
 
 }
