+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/entities/TaskTag.java
@@ -1,4 +1,4 @@
-package com.taskmanager.taskmanagerapi.dto;
+package com.taskmanager.taskmanagerapi.entities;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
