+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/repositories/ProjectRepository.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.repositories;
 
-import com.taskmanager.taskmanagerapi.dto.Project;
+import com.taskmanager.taskmanagerapi.entities.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
