+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/repositories/ProjectTaskRepository.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.repositories;
 
-import com.taskmanager.taskmanagerapi.dto.ProjectTask;
+import com.taskmanager.taskmanagerapi.entities.ProjectTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
