+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/repositories/ReportRepository.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.repositories;
 
-import com.taskmanager.taskmanagerapi.dto.Report;
+import com.taskmanager.taskmanagerapi.entities.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
