+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/repositories/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.repositories;
 
-import com.taskmanager.taskmanagerapi.dto.Role;
+import com.taskmanager.taskmanagerapi.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
