+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/repositories/TaskRepository.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.repositories;
 
-import com.taskmanager.taskmanagerapi.dto.Task;
+import com.taskmanager.taskmanagerapi.entities.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
