+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/repositories/TaskTagRepository.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.repositories;
 
-import com.taskmanager.taskmanagerapi.dto.TaskTag;
+import com.taskmanager.taskmanagerapi.entities.TaskTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
