+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/repositories/UserRepository.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.repositories;
 
-import com.taskmanager.taskmanagerapi.dto.User;
+import com.taskmanager.taskmanagerapi.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
