+++ task-manager-api/src/main/java/com/taskmanager/taskmanagerapi/repositories/UserRoleRepository.java
@@ -1,6 +1,6 @@
 package com.taskmanager.taskmanagerapi.repositories;
 
-import com.taskmanager.taskmanagerapi.dto.UserRole;
+import com.taskmanager.taskmanagerapi.entities.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
